Add support for multiple palettes
Minecraft shipwrecks use multiple palettes, allowing a single structure file to contain various schemes depending on the location the structure is placed.

All of the code supporting this exists, except for the actual loading of multiple palettes. Structure.setSeed() is called using the chunk coords, and Structure.getObject() as well, prior to rendering the final object into the world.

Attempting to load an NBT file with multiple palettes causes a NullPointerException and provides no information about the error. Now, loading an NBT with multiple palettes is supported and works as expected.